array_and_slices/sum_all_test.go
package arrayandslices

import (
	"reflect"
	"testing"
)

func TestSumAll(t *testing.T) {

	t.Run("called two slice and sum", func(t *testing.T) {
		got := SumAll([]int{1, 2}, []int{0, 9})
		expected := []int{3, 9}

		if !reflect.DeepEqual(got, expected) {
			t.Errorf("got '%v' want '%v'", got, expected)
		}
	})
}

func TestSumAllTails(t *testing.T) {

	checkSums := func(t testing.TB, got, expected []int) {
		t.Helper()
		if !reflect.DeepEqual(got, expected) {
			t.Errorf("got '%v' want '%v'", got, expected)
		}
	}
	t.Run("make sum all slice tails", func(t *testing.T) {
		got := SumAllTails([]int{1, 2}, []int{0, 9})
		expected := []int{2, 9}

		checkSums(t, got, expected)
	})

	t.Run("safely sum empty slice tails", func(t *testing.T) {
		got := SumAllTails([]int{}, []int{3, 4, 5})
		expected := []int{0, 9}

		checkSums(t, got, expected)
	})
}
